dev/reset-conda-env.bat works finally
The problem was with the Windows AppData\Cache, which poetry tries to use.
Looks like the cache is not consistent in time and poetry fails to find dependencies
which it expects to be there. Established special cache for poetry, which can be shared
between multiple projects. Probably the same cache is to be set for pip.